feat: converted everything to cli
