+++ src/repror/app.py
@@ -1,0 +1,7 @@
+from repror.cli import cli
+
+# This is the entry point for the CLI
+app = cli.app
+
+if __name__ == "__main__":
+    app()
