+++ src/repror/cli/generate_recipes.py
@@ -2,7 +2,8 @@ import json
 from repror.internals import conf
 
 
-if __name__ == "__main__":
+def generate_recipes():
+    """Generate list of recipes from the configuration file."""
     # Prepare the matrix
     recipe_list = []
